+++ archiva-modules/archiva-base/archiva-proxy/src/test/java/org/apache/archiva/metadata/repository/TestRepositorySessionFactory.java
@@ -31,7 +31,7 @@ public class TestRepositorySessionFactory
     @Override
     public RepositorySession createSession()
     {
-        return new RepositorySession( new TestMetadataRepository(), resolver );
+        return new RepositorySession( new AbstractMetadataRepository() {}, resolver );
     }
 
     public void setResolver( MetadataResolver resolver )
